feat: Change price class

+++ src/main/java/com/xaut/shop/domain/entity/Order.java
@@ -10,7 +10,7 @@ public class Order {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
 
-    private float totalPrice;
+    private double totalPrice;
     private String status;
     private Date createTime;
     private Date finishTime;
@@ -30,11 +30,11 @@ public class Order {
         this.id = id;
     }
 
-    public float getTotalPrice() {
+    public double getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(float totalPrice) {
+    public void setTotalPrice(double totalPrice) {
         this.totalPrice = totalPrice;
     }
